Change to snippet format
