+++ package.json
@@ -11,8 +11,5 @@
     "license": "ISC",
     "devDependencies": {
         "prettier": "^1.17.0"
-    },
-    "dependencies": {
-        "serverless-cloudflare-workers": "^1.2.0"
     }
 }
